Updated Perimeter detection logic
Adds clearer logic for when commands are cut short, for example:

A command to move 10000m from square 1 facing South will move to the furthest row ,100/square 9901, this is cut short as there are only 100x100 squares, we now display a clear message in the Mission Log to say '(Move cut short by perimeter, moved 99m)'.

We are now tracking the requested input of 10000m vs the actual distance travelled which has to be cut short to 99m and we're then clearly displaying this in the Mission Log.